Commit code 20171008

+++ mtv_management_leave/Models/Entity/UserSeniority.cs
@@ -10,18 +10,18 @@ namespace mtv_management_leave.Models.Entity
         public int Id { get; set; }
         public int Uid { get; set; }
         public int? Year { get; set; }
-        public byte? Month1 { get; set; }
-        public byte? Month2 { get; set; }
-        public byte? Month3 { get; set; }
-        public byte? Month4 { get; set; }
-        public byte? Month5 { get; set; }
-        public byte? Month6 { get; set; }
-        public byte? Month7 { get; set; }
-        public byte? Month8 { get; set; }
-        public byte? Month9 { get; set; }
-        public byte? Month10 { get; set; }
-        public byte? Month11 { get; set; }
-        public byte? Month12 { get; set; }
+        public int? Month1 { get; set; }
+        public int? Month2 { get; set; }
+        public int? Month3 { get; set; }
+        public int? Month4 { get; set; }
+        public int? Month5 { get; set; }
+        public int? Month6 { get; set; }
+        public int? Month7 { get; set; }
+        public int? Month8 { get; set; }
+        public int? Month9 { get; set; }
+        public int? Month10 { get; set; }
+        public int? Month11 { get; set; }
+        public int? Month12 { get; set; }
         [ForeignKey("Uid")]
         public virtual UserInfo UserInfo { get; set; }
     }
